+++ tasks/mozfest_backups/guidebook_rollback.py
@@ -1,0 +1,10 @@
+#  TODO:
+# Rollback (separate file + made to be executed locally): do a diff on what's actually on guidebook (redo a dump of everything) and last backup.
+# Check if tracks, location, session are different.
+# Do a intersection set with session IDs + diff
+# Check what's new (new ids)
+# Check what was removed (ids that are not in the backup)
+# create different jsons with all changes (session updates, deletes and additions)
+# do requests to guidebook
+
+# Need a CLI for people to run the rollback!
